moving TESTAR GUI into settingdialog package
