+++ testar/src/org/testar/extendedsettings/ExampleSetting.java
@@ -28,14 +28,28 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************************************/
 
-package org.testar.settings;
+package org.testar.extendedsettings;
 
-public interface IExtendedSettingDefaultValue<T> {
-    /**
-     * Functor to create an initialized object T with default values.
-     * Should not be implemented by specializations of {@link ExtendedSettingFile}.
-     *
-     * @return Object T initialized with default values.
-     */
-    ExtendedSettingBase<T> CreateDefault();
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlRootElement;
+
+@XmlRootElement
+@XmlAccessorType(XmlAccessType.FIELD)
+public class ExampleSetting extends ExtendedSettingBase<ExampleSetting> {
+    public String test;
+
+    public static ExampleSetting CreateDefault() {
+        ExampleSetting instance = new ExampleSetting();
+        instance.test = "Hello";
+        return instance;
+    }
+
+    @Override
+    public int compareTo(ExampleSetting other) {
+        if (test.contentEquals(other.test)){
+            return 0;
+        }
+        return -1;
+    }
 }
