+++ testar/src/org/testar/extendedsettings/ExtendedSettingContainer.java
@@ -28,28 +28,36 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************************************/
 
-package org.testar.settings;
+package org.testar.extendedsettings;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
+import java.util.Observable;
+import java.util.Observer;
 
-@XmlRootElement
-@XmlAccessorType(XmlAccessType.FIELD)
-public class ExampleSetting extends ExtendedSettingBase<ExampleSetting> {
-    public String test;
+public class ExtendedSettingContainer<S extends ExtendedSettingBase<S>> implements Observer {
+    // Access to the resource file which also holds the old value in case we need to update the settings.
+    private final ExtendedSettingFile _file;
+    // The actual settings.
+    private final ExtendedSettingBase<S> _setting;
 
-    public static ExampleSetting CreateDefault() {
-        ExampleSetting instance = new ExampleSetting();
-        instance.test = "Hello";
-        return instance;
+    public ExtendedSettingContainer(ExtendedSettingFile file, Class<S> clazz,
+                                    IExtendedSettingDefaultValue<S> defaultFunctor) {
+        _file = file;
+        _setting = _file.load(clazz, defaultFunctor);
+        _setting.addObserver(this);
+    }
+
+    /**
+     * Get the settings.
+     * @return The actual settings.
+     */
+    ExtendedSettingBase<S> GetSettings() {
+        return _setting;
     }
 
     @Override
-    public int compareTo(ExampleSetting other) {
-        if (test.contentEquals(other.test)){
-            return 0;
+    public void update(Observable o, Object arg) {
+        if (o.equals(_setting)) {
+            _file.save(_setting);
         }
-        return -1;
     }
 }
