+++ testar/src/org/testar/settingsdialog/SettingsPanel.java
@@ -28,19 +28,33 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************************************/
 
-package org.testar.settings;
+package org.testar.settingsdialog;
 
+import org.testar.monkey.Settings;
 
-import java.io.Serializable;
-import java.util.Observable;
+import javax.swing.*;
 
-public abstract class ExtendedSettingBase<T> extends Observable implements IExtendedSetting, Comparable<T>, Serializable {
+/**
+ * Abstract class for all common functionality for a SettingsPanel.
+ */
+public abstract class SettingsPanel extends JPanel {
     /**
-     * Notify the {@link IExtendedSettingContainer} that the specialization of this class needs to be saved.
+     * Populate the fields from Settings structure.
+     *
+     * @param settings The settings to load.
      */
-    @Override
-    public void Save() {
-        setChanged();
-        notifyObservers();
+    public abstract void populateFrom(final Settings settings);
+
+    /**
+     * Retrieve information from the GUI.
+     *
+     * @param settings reference to the object where the settings will be stored.
+     */
+    public abstract void extractInformation(final Settings settings);
+
+    /**
+     * Validate that the settings are valid.
+     */
+    public void checkSettings() {
     }
 }
