+++ testar/src/org/testar/settingsdialog/dialog/FilterPanel.java
@@ -28,12 +28,12 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************************************/
 
-package org.testar.monkey.dialog;
+package org.testar.settingsdialog.dialog;
 
 import org.apache.commons.lang3.StringUtils;
 import org.testar.monkey.ConfigTags;
 import org.testar.monkey.Settings;
-import org.testar.monkey.SettingsPanel;
+import org.testar.settingsdialog.SettingsPanel;
 
 import javax.swing.*;
