+++ testar/src/org/testar/settingsdialog/dialog/TimingPanel.java
@@ -29,17 +29,13 @@
 *******************************************************************************************************/
 
 
-package org.testar.monkey.dialog;
+package org.testar.settingsdialog.dialog;
 
 import org.testar.monkey.ConfigTags;
 import org.testar.monkey.Settings;
-import org.testar.monkey.SettingsPanel;
+import org.testar.settingsdialog.SettingsPanel;
 
 import javax.swing.*;
-
-import static org.testar.monkey.dialog.ToolTipTexts.*;
-import static org.testar.monkey.dialog.ToolTipTexts.maxTestTimeTTT;
-import static org.testar.monkey.dialog.ToolTipTexts.useRecordedTTT;
 
 public class TimingPanel extends SettingsPanel {
 
@@ -62,81 +58,81 @@ public class TimingPanel extends SettingsPanel {
     spnActionDuration = new JSpinner();
     spnActionDuration.setBounds(143, 14, 100, 27);
     spnActionDuration.setModel(new SpinnerNumberModel(0.0d, 0.0d, null, 0.1d));
-    spnActionDuration.setToolTipText(actionDurationTTT);
+    spnActionDuration.setToolTipText(ToolTipTexts.actionDurationTTT);
     add(spnActionDuration);
 
     spnActionWaitTime = new JSpinner();
     spnActionWaitTime.setBounds(143, 52, 100, 27);
     spnActionWaitTime.setModel(new SpinnerNumberModel(0.0d, 0.0d, null, 0.1d));
-    spnActionWaitTime.setToolTipText(actionWaitTimeTTT);
+    spnActionWaitTime.setToolTipText(ToolTipTexts.actionWaitTimeTTT);
     add(spnActionWaitTime);
 
     spnSutStartupTime = new JSpinner();
     spnSutStartupTime.setBounds(143, 90, 100, 27);
     spnSutStartupTime.setModel(new SpinnerNumberModel(0.0d, 0.0d, null, 1.0d));
-    spnSutStartupTime.setToolTipText(sutStartupTimeTTT);
+    spnSutStartupTime.setToolTipText(ToolTipTexts.sutStartupTimeTTT);
     add(spnSutStartupTime);
 
     spnMaxTime = new JSpinner();
     spnMaxTime.setBounds(143, 128, 100, 31);
     spnMaxTime.setModel(new SpinnerNumberModel(0.0d, 0.0d, null, 1.0d));
-    spnMaxTime.setToolTipText(maxTestTimeTTT);
+    spnMaxTime.setToolTipText(ToolTipTexts.maxTestTimeTTT);
     add(spnMaxTime);
 
     checkUseRecordedTimes = new JCheckBox();
     checkUseRecordedTimes.setBounds(271, 177, 21, 21);
-    checkUseRecordedTimes.setToolTipText(useRecordedTTT);
+    checkUseRecordedTimes.setToolTipText(ToolTipTexts.useRecordedTTT);
     add(checkUseRecordedTimes);
   }
 
   private void addTimingLabels() {
     JLabel jLabel2 = new JLabel("Action Duration:");
     jLabel2.setBounds(10, 14, 130, 14);
-    jLabel2.setToolTipText(actionDurationTTT);
+    jLabel2.setToolTipText(ToolTipTexts.actionDurationTTT);
     add(jLabel2);
 
     JLabel jLabel3 = new JLabel("seconds");
     jLabel3.setBounds(256, 17, 52, 14);
-    jLabel3.setToolTipText(actionDurationTTT);
+    jLabel3.setToolTipText(ToolTipTexts.actionDurationTTT);
     add(jLabel3);
 
 
     JLabel jLabel4 = new JLabel();
     jLabel4.setBounds(10, 52, 130, 14);
     jLabel4.setText("Action Wait Time:");
-    jLabel4.setToolTipText(actionWaitTimeTTT);
+    jLabel4.setToolTipText(ToolTipTexts.actionWaitTimeTTT);
     add(jLabel4);
 
     JLabel jLabel5 = new JLabel("seconds");
     jLabel5.setBounds(256, 55, 52, 14);
-    jLabel5.setToolTipText(actionWaitTimeTTT);
+    jLabel5.setToolTipText(ToolTipTexts.actionWaitTimeTTT);
     add(jLabel5);
 
 
     JLabel jLabel7 = new JLabel("seconds");
     jLabel7.setBounds(256, 93, 52, 14);
-    jLabel7.setToolTipText(sutStartupTimeTTT);
+    jLabel7.setToolTipText(ToolTipTexts.sutStartupTimeTTT);
     add(jLabel7);
 
     JLabel jLabel6 = new JLabel("SUT Startup Time:");
     jLabel6.setBounds(10, 90, 130, 14);
-    jLabel6.setToolTipText(sutStartupTimeTTT);
+    jLabel6.setToolTipText(ToolTipTexts.sutStartupTimeTTT);
     add(jLabel6);
 
 
     JLabel jLabel22 = new JLabel("Max. Test Time:");
     jLabel22.setBounds(10, 128, 130, 14);
-    jLabel22.setToolTipText(maxTestTimeTTT);
+    jLabel22.setToolTipText(ToolTipTexts.maxTestTimeTTT);
     add(jLabel22);
 
     JLabel jLabel23 = new JLabel("seconds");
     jLabel23.setBounds(256, 131, 52, 14);
-    jLabel23.setToolTipText(maxTestTimeTTT);
+    jLabel23.setToolTipText(ToolTipTexts.maxTestTimeTTT);
     add(jLabel23);
 
     JLabel jLabel24 = new JLabel("Use Recorded Action Timing during Replay:");
     jLabel24.setBounds(10, 177, 255, 14);
-    jLabel24.setToolTipText(useRecordedTTT);
+    jLabel24.setToolTipText(ToolTipTexts.useRecordedTTT);
     add(jLabel24);
   }
