+++ testar/test/org/testar/settings/TestSetting.java
@@ -30,10 +30,11 @@
 
 package org.testar.settings;
 
+import org.testar.extendedsettings.ExtendedSettingBase;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
-import java.io.Serializable;
 
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
